Terminando con la gestion de personas( Guardar, Modificar, Dar de baja)y todo lo relacionado a ello incluido el manejo de imagenes para Lectura y escritura de Archivos de imagen desde la BDD

+++ ProyectoFinal-Antecedentes/src/ControlAdminDatos/RolDao.java
@@ -77,12 +77,12 @@ public class RolDao implements Dao<Rol>{
     }
 
     @Override
-    public void destroy(Integer id) {
+    public void destroy(Long id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public Rol find(Integer id) {
+    public Rol find(Long id) {
         Rol rol = null;
         try {
             //Cargar la lista de cuentas
